+++ src/app/services/redux-constants.ts
@@ -10,9 +10,9 @@ export const SET_SAMPLE_QUERY_SUCCESS = 'sampleQuery/setSampleQuery';
 export const SET_GRAPH_EXPLORER_MODE_SUCCESS = 'graphExplorerMode/setGraphExplorerMode';
 export const PROFILE_REQUEST_SUCCESS = 'profile/getProfileInfo/success';
 export const PROFILE_REQUEST_ERROR = 'profile/getProfileInfo/error';
-export const GET_SNIPPET_SUCCESS = 'GET_SNIPPET_SUCCESS';
-export const GET_SNIPPET_ERROR = 'GET_SNIPPET_ERROR';
-export const GET_SNIPPET_PENDING = 'GET_SNIPPET_PENDING';
+export const GET_SNIPPET_SUCCESS = 'snippet/getSnippet/fulfilled';
+export const GET_SNIPPET_ERROR = 'snippet/getSnippet/rejected';
+export const GET_SNIPPET_PENDING = 'snippet/getSnippet/pending';
 export const REMOVE_HISTORY_ITEM_SUCCESS = 'history/removeHistoryItem';
 export const ADD_HISTORY_ITEM_SUCCESS = 'history/addHistoryItem';
 export const GET_HISTORY_ITEMS_SUCCESS = 'GET_HISTORY_ITEMS_SUCCESS';
